Ajustando lógica de remoção de transportes

+++ src/controllers/trip/removeTrip.controller.js
@@ -2,17 +2,27 @@ const tripModel = require("../../models/trip.model");
 
 async function removeTrip(tripId) {
     try {
-        const deletedTrip = await tripModel.findByIdAndDelete(tripId);
-        if (!deletedTrip) {
+        const trip = await tripModel.findById(tripId);
+        if (!trip) {
             return {
                 status: 200,
                 message: "Trip not found"
             };
+        } else if (trip.patients.length) {
+            return {
+                status: 200,
+                message: "trip has patients"
+            }
+        } else {
+            await tripModel.findByIdAndDelete(tripId)
+                .then(() => {
+                    return {
+                        status: 200,
+                        message: "Trip removed successfully"
+                    };
+                })
         }
-        return {
-            status: 200,
-            message: "Trip removed successfully"
-        };
+       
     } catch (error) {
         return {
             status: 200,
